Avoid mixed double float operations.
On my machine moving over to float to avoid the mixed double/float operation slowed things
down even more.  We may want to revisit this later.  Diffs were minor and had about
10 percent speed up on interpolation with this change.

+++ gs/base/siscale.c
@@ -30,7 +30,7 @@
 
 /* Auxiliary structures. */
 typedef struct {
-    float weight;               /* float or scaled fraction */
+    double weight;               /* float or scaled fraction */
 } CONTRIB;
 
 typedef struct {
